requirements.txt
pandas==2.2.2
numpy==2.0.2
PuLP==2.9.0
requests==2.32.3
PyYAML==6.0.2